requirements.txt
Django==4.2.14
djangorestframework==3.15.2
django-mailer>=2.3.2
gspread==5.9.0
oauth2client>=4.0.0
psycopg2-binary==2.9.6
wagtail==5.0.5
openpyxl~=3.1.1
crypto~=1.4.1
gunicorn==22.0.0
whitenoise==6.6.0
markdown==3.7.0
# Testing
pytest==8.2.2
pytest-django==4.8.0